Extracted janus tunnel and process handler into a separate module
Datachannel for faster device-browser communication is no longer a pro feature.
